feat: move models into the core
