+++ bin/Cargo.toml
@@ -9,12 +9,12 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0"
 env_logger = "0.7.1"
-http = "0.2"
+#http = "0.2"
 log = "0.4.0"
 rustacles-brokers = { git = "https://github.com/spec-tacles/rustacles", branch = "feat/rpc" }
 serde = "1.0"
 serde_json = "1.0"
-serde_repr = "0.1"
+#serde_repr = "0.1"
 spectacles-proxy = { path = "../core", features = ["redis-ratelimiter"] }
 tokio = { version = "0.2", features = ["rt-core", "rt-threaded", "macros", "signal"] }
 toml = "0.5"
