+++ bin/src/main.rs
@@ -16,13 +16,11 @@ use uriparse::Scheme;
 
 mod client;
 mod config;
-mod models;
 #[cfg(test)]
 mod test;
 
 pub use client::*;
 pub use config::*;
-pub use models::*;
 
 #[tokio::main]
 async fn main() {
