+++ core/Cargo.toml
@@ -9,9 +9,11 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0"
 env_logger = "0.7"
+http = "0.2"
 log = "0.4"
 serde = "1.0"
 serde_json = "1.0"
+serde_repr = "0.1"
 tokio = { version = "0.2", features = ["blocking", "rt-core", "stream", "sync"] }
 uriparse = "0.6"
